Updated effingpropa and indelboy to operate per sample rather than per directory and renamed.
