+++ .gitignore
@@ -2,3 +2,5 @@
 .DS_Store
 __pycache__
 *.lprof
+.snakemake/*
+complete.txt
